requirements.txt
flask==3.1.0
flask-apscheduler==1.13.1
flask-mysqldb==2.0.0
google-cloud-secret-manager==2.23.2
numpy==2.2.4
pandas==2.2.3
pymysql==1.1.1
requests==2.32.3
werkzeug==3.1.3